// src/types/rows/OrcamentoCabecalhoPrevistoPorMes.ts
export interface OrcamentoCabecalhoPrevistoPorMes {
  anomes: string;
  faturamento_bruto_prev: number;
  imposto_prev: number;
  cmv_prev: number;
  dif_inv_prev: number;
  lucro_bruto_prev: number;
  despesas_previsto: number;
  perdas_previsto: number;
  receita_extra_financ_prev: number;
  receita_extra_comercial_prev: number;
  receita_extra_mkt_prev: number;
  lucro_operacional_previsto: number;
  irpj_prev: number;
  csll_prev: number;
  lucro_liquido_prev: number;
}